Move input into a proper file
